Updates the game logic
Changes around the logic to accept an array of cells or to get the
current state of the game board that is rendered if that array is not
supplied.

This will allow me to determine a game outcome based on the way it was
stored in the database OR get the outcome of the current game before
saving it to the database.